requirements.txt
fastapi~=0.115.13
sqlalchemy~=2.0.41
pymysql
bcrypt~=4.3.0
PyJWT~=2.10.1
pydantic~=2.11.7
passlib~=1.7.4